Implement MuSig on Ristretto (#162)
* Implement MuSig on Ristretto

This implementation is implemented on two layers: A public API layer (MuSig), and a private Finite State Machine. The purpose of doing it like this has a few advantages from a security perspective:

All private data is completely inaccessible to users
Sensitive temporary data is dropped as the state machine progresses.
Any invalid actions immediately causes the MuSig state to become invalid, and all sensitive data is immediately destroyed.
The state transition set is highly constrained. As a result it is impossible to rewind the MuSig ceremony in order to execute a replay attack; The Message that is being signed can be set exactly once.

MuSig takes ownership of the public nonces, mitigating the risk of exposing private keys due to nonce reuse (it's not impossible, but a user must explicitly clone() a public nonce, which should give him a hint that he's doing something dangerous).