Refactor HomomorphicCommitment
HomomorphicCommitment is refactored to abstract away the static reference to the Commitment
base without compromising on the ability to build commitments with different generator bases.

The explicit dependency on the `Scalar` from the Dalek library is also removed, which removes
the leaky abstraction that was present in the existing implementation.

The references to `base` that were being carried in `Transaction` structs can also now be
removed.