Remove Copy trait for SecretKey
The `Copy` trait is probably just a bad idea to have on SecretKey.
This is removed, but required some refactoring. This commit
refactors the `crypto` crate so that `cargo test -p crypto`
passes. The *good* news is that the test tat checks that
secretkey memory is cleared out passes now.

There's a separate PR that's working on `Transaction`, so I've left off
refactoring that for now, since it will amount ot double work.