Implement the transaction builder for the Sender (#239)
* Implement the transaction builder for the Sender

The SenderTransactionInitializer is a Builder that helps set up the initial state for the Sender party of a new transaction
Typically you don't instantiate this object directly. Rather use
```
SenderTransactionProtocol::new(1);
```
which returns an instance of this builder. Once all the sender's information has been added via the builder
methods, you can call `build()` which will return the ProtocolState instance

* Make a few changes based on review comments