Provide a bulletproof range proof implementation (#240)
* Provide a bulletproof range proof implementation

This PR closes #61.

We present a bulletproof range proof based on the Dalek library implementation
(see https://doc.dalek.rs/bulletproofs/index.html). This implementation is a
thin wrapper around the Dalek library so that a unified API is presented to the
Tari domain layer. Should another underlying library be preferred, it is relatively
simple to swap them out.

Only sinlge proofs are supported in the API at this time, but it is easy to extend
this to aggregated proofs in the future.

* Cargo lock