Sort kernels
Kernel sorting was temp removed from the code, but now an efficient ordering mechanism
has been implemented on `SchnorrSignature` that allows Ord to be derived on TransactionKernel.
This ordering does not require 2n.ln(n) hash operation, but in most cases zero, and occasionally
a few `to_hex` conversions if signature public nonces are equal for whatever reason.