Split Hex out from ByteArray (#264)
* Split Hex out from ByteArray

ByteArray cannot be implemnted on types without an intrinsic single binary structure
(e.g. SchnorrSignature, which has two fields, and so can implement `as_bytes() -> &[u8]`.
But these types _can_ still implment `to_hex` since a new string is created to hold the
data.

This commit splits out the hex conversion functions in to a separate trait and provides
a blanket implmentation for `ByteArray` so that nonthing needs to change in existing
code (except for an import).