Abstraction of zmq::proxy (#320)
Small abstraction of zmq proxy which is used in inbound message service. This may also be needed elsewhere (e.g. ControlService). The proxy requires Connection to be able to bind with an Inbound direction, so I've added another property to connections which allows you to optionally select the way the socket establishes connections (Bind, Connect or Auto(default))