Fix broken tests in release mode
It looks like Rust can already be reusing memory by the time the address is checked after being zeroed.
e.g. see this code
```
use rand::rngs::OsRng;
use rand::Rng;
use std::slice;
use std::mem;
use std::ptr;

pub struct Zeroed([u8; 32]);

impl Drop for Zeroed {
    fn drop(&mut self) {
        let size = mem::size_of_val(self);
        unsafe {
            let ptr = self as *mut Self;
            ptr::write_bytes(ptr as *mut u8, 0, size);
        }
        println!("In Drop: {:?}", self);
    }
}

fn main() {
    let zero = &vec![0u8; 32][..];
    let mut rng = OsRng::new().unwrap();
    let ptr;
    {
        let mut k = [0u8;32];
        rng.fill(&mut k);
        let foo = Zeroed(k);
        println!("Value: {:?}", foo);
        ptr = foo.0.as_ptr();
        println!("Address before: {:?}", ptr);
        unsafe { assert_eq!(slice::from_raw_parts(ptr, 32), k); }
    }
    unsafe {
        assert_eq!(slice::from_raw_parts(ptr, 32), zero);
        println!("Address after: {:?}", ptr);
        println!("Value after: {:?}", slice::from_raw_parts(ptr, 32));    }
}
```

It passes in Debug, but typically fails in release mode. But the memory is always correclty zeroed.
Therefore, we opt to ignore the zeroing checks in release compiles

Some tests check the value of a call to `System.now()`. In release mode, this can be so fast that the clock doesn't tick,
and tests that rely on a tick will fail. Our workaround is to add a tiny delay so that the timestamps do change when the
system time is read.