Control service, refactor dispatcher, partial refactor IMS (#329)
Added a control service which listens for and handles incoming connections from peers wanting to initiate a PeerConnection. The ControlService follows similar patterns to PeerConnection, with a control service starting a worker which can be controlled by sending messages to that worker thread.

* Dispatcher refactor:

The dispatcher "route type"/"key"/"dispatch type" can be any type that is Eq + Hash
Added the concept of a resolver. This is to allow custom mappings for messages to a particular route.
Added catch_all, which can be optionally used to handle all messages which don't have a corresponding route.
The dispatcher in IMS and control service has been aliased with concrete types to match their use-case.

* IMS refactor:
  * Use the new dispatcher
  * Allow tests to pass with the new concrete dispatcher

TODO: update tests to test the Handle, Forward, Discard conditions once they are implemented in the InboundMessageServiceResolver

* Messages refactor:
  * Moved message envelope into tari_comms::messages
  * Added structs for Message and MessageHeader (as per RFC-0172)
  * Made extensive use of the default impl of MessageFormat and removed try_intos (good mini-debate, but I feel that std::convert::TryInto should not be used for deserialization - just a feeling though, can't back that up with a reason :P - maybe just a clarity thing - either way, the code ends up being similar to the default MessageFormat impl)
* Still to be done:
  * Implement comms-level/default handlers for control service (namely establish_connection handler)
  * Once an establish connection has been completed/failed, find some way to disconnect (forcibly) the peer from the control port to release resources.
  * Message decryption
  * Peer banning, flood protection, heuristics etc.

* Ignore all .log files

This will ignore all empty log directories.
Logs currently appear (with some changes to the path in initialize_logging)
in the subdir that the binary is run from, not just common.
This will be fixed when issue #328 is fixed.

* Added control service + refactored dispatcher + other refactors

* Fixed compile errors in IMS and OMS resulting from refactor

* Moved Frame and FrameSet to message