Remove port range requests (#354)
The approach of scanning ports in a range to find ephemeral ports to make connections with was
causing problems on two counts:
1. Rust runs tests in parallel, and scanning from 10000 for each test meant that there was
  a race condition when two tests check a port, see that it's free and then one gets it and the other
  does not.
2. Race conditions between `TcpListener::bind` and `zmq::connect`. Basically to see if a port is free
   we'd bind to it. If that was succesful, drop the bind and then give amq the port. If the OS didn't
   release the port in time, the zmq::connect call would fail.

We take a new approach to this here. Firstly, outbound connections must have their port specified
(This is typically done in config files). Inbound connections can have a random port selected
by using zero as the port, i.e. `0.0.0.0:0`. This delegates port selection to the OS and presumably
it does a better job of handling these race conditions.

After the connection is made, you can see the actual port that was used by calling
`conn.get_connected_address()`

This hopefully addresses the test flakiness and presents a more robust approach to port selection.
As a result all the port range and port selection logic has been removed as well.