Added OutboundMessagePool (#355)
The Outbound Message Pool (OMP) fields requests to send OutboundMessages from the various OutboundMessageService's (OMS's) in the system. OMS's assemble the OutboundMessages to be sent and then write them to the OMP's inbound InProc address which acts as a queue. The OMP spins up a series of workers to which the messages are fairly dealt and each worker will attempt to send the message, if it has failed it will requeue the message to the OMP queue to be attempted again in the future.

Currently the OMP is almost done, all the threading, worker logic and queuing is done. it is just waiting on the ConnectionService to be completed to actually attempt to make connection and attempt to transmit the message.