This is a not-quite drop-in replacement for LMDBStore.
It's presented in parallel so that users can migrate at will.

This version is much more ergonomic and presents a less leaky abstraction than the current implementation.

Key changes:

There's no more "current db". You can get a handle to a DB and work with it for as long as you like.

* no more get_raw. Everything is abstracted away to get<T> and insert<T> where T: (De)Serialize.
* iterators! (almost). the for_each function offers the same functionality and works just like Vec::for_each.
* len() gives the size of the DB (table)
* log_info() can be used to periodically dump useful stats to the log.

The docstring outlines the journey to reaching this design and the tradeoffs made wrt zero copy and being opinionated about serialisation.