Refactor RFC layout
The Tari Network architecture has started to solidify and this commit
reflects the maturation of that thinking.

The RFCs are refactored and consolidated  into three main chapters (with some auxiliary chapters)
that talk to the three main layers of the Tari Netowrk: Base layer, Payment
channels, and Digital Assets Network.