NodeId de/serialized to/from hex (#415)
* Description

NodeId will now be de/serialized to/from hex, by implementing ByteArray. I moved the serialize_to_hex function from tari_crypto to tari_utilities as the same function can be used for serialization in that crate. I was not able to do the same for deserialization.

* Motivation and Context

NodeId was serializing to json as an array while public key was a hex string. This change makes the serialization of peers consistent.