PingPong example for tari_p2p (#421)
A relatively small pingpong example using the tari_p2p and tari_comms libraries.
Testing on comms from two different machines (mac and linux) seemed to
show that there is some issue with peer connection establishment which still
needs to be worked out. I've PR'd this example for everyone to play with
in the meantime.