Custom serialisation for Pub- and SecretKey
Implment custom serialisation for RistrettoPublicKey and RistrettoSecretKey
that uses the `is_human_readable` function to decide whether to use Hex
or Binary for encoding.

This required a change in the MessageFormat binary encosgin scheme
since RMP-serde sets `human_readable` to true (actually it neglects
to set it to false). So I switched the binary format to bincode,
which is substantially faster anyway.

This _probably_ makes the `pubkey_from_hex` function obsolete, but will
check this out in a separate PR