Remove logging from tests
The logs in CI are exceeding circle ci's maximum number of log lines
and are cluttering test output. When debugging, you can add
`simple_logger::init()` to the test you are debugging. This will
also emit a warning which would remind you to remove it afterwards.

In future, we may want logs in CI tests to write to a file and which can be
later inspected as an [artifact](https://circleci.com/docs/2.0/artifacts/).