Improved and secure connection establishment protocol
Steps to establish a connection as follows (Node A connecting to node B):

1) Node A sends a RequestConnection message to the node B's control port
2) If node B accepts the request, it opens an inbound `PeerConnection`
and returns the details of the connection in an encrypted message back
through the control port.
3) Node A uses those details to create an outbound `PeerConnection` to the given `NetAddress`

Previously Node A could request that Node B connects to an address which
meant it could request Node B (and many other nodes) to spam/flood a particular endpoint.
With this new scheme, node A has no control over where node B connects.
If node B can at worst tell node A to connect to a random target (would
fail) but node A would have had to initiate the request for that to
happen. There's no way for a node to tell the network to spam others.