Remove some sensitive functions from BlockBuilder
- We can't pass privkey to a function like `create_coinbase`

Private keys will often come from wallets and be passed to a node via, e.g. gRPC.
Private should never be transmitted where we can help it.
in this case, wallets are perfectly capable of creating and signing the coinbase
UTXO and kernel.

The better route is to build the UTXO in the wallet and use `with_coinbase_utxo` to add it to
the block.

- DRY up test code
SimpleBlockChain had several instances of repeated code. Clean those up so that
there's no repetition.