Remove some sensitive functions from BlockBuilder (#627)
* Move transaction validation code from `Transaction` to `AggregateBody`

Nearly all the transaction validation code currently in `Transaction`
involves references to `self.body.some_field`. This suggests that
it should really live in `AggregateBody`.

As it stands _block_ validation code has to do some yucky copying
to get at the same code (blocks have an AggregateBody, not a
Transaction).

This PR moves all the validation code into `AggregateBody` which
both cleans up the code smell, and allows a more efficient and
ergonomic call through to validation code when validating `Block`s

* Remove some sensitive functions from BlockBuilder

- We can't pass privkey to a function like `create_coinbase`

Private keys will often come from wallets and be passed to a node via, e.g. gRPC.
Private should never be transmitted where we can help it.
in this case, wallets are perfectly capable of creating and signing the coinbase
UTXO and kernel.

The better route is to build the UTXO in the wallet and use `with_coinbase_utxo` to add it to
the block.

- DRY up test code
SimpleBlockChain had several instances of repeated code. Clean those up so that
there's no repetition.

* Update test chain data