Update RFC-0111
After some consideration, we're changing the message-passing proposal
from using channels to tokio-based futures. Tokio's default task executor
provides multi-threaded work-stealing work queues and CPU-bound worker
threads out of the box. This is a good fit for the type of work that base nodes must perform. In addition, the
[Tower project](https://github.com/tower-rs) provides a set of traits and
middleware that will be very useful in Tari services, and so it is
recommended to follow the Services pattern as used by that project.