Refactor KeyValStore trait to be less restrictive (#662)
As it stands, the `KeyValStore` trait has some pretty restrictive constraints
on it, inlcuding the very ugly dependency on an LMDB type.

There's no reason to impose these constraints, actually. Every implementation
is free to impose additional constraints to limit its applicability.

The main change of this PR is to remove the bounds on the `K` and `V`
generics in the trait, and shift them up into the implementations.

This allows us to use `KeyValStore` as a backend trait in e.g.
MerkleMountainRange without having to carry things like `AsLMDBBytes`
into the MMR crate :O

It also lets us massively simply the `HMapDatabase` type (no serialisation
required!)

So, two big wins!!

But..it makes things a little messy with LMDB.
There's a conflict between where the Generic specifications (K,V) live.
It makes sense to have them at the trait level for the trait; but I couldn't
find a way to put them on the struct level for LMDBDatabase; primarily
because `LMDBStore` holds a map of `LMDBDatabase` instances, each with
their own, potentially different `K` and `V` types - i.e. `LMDBDatabase` can't
be `LMDBDatabase<K,V>`.

So my workaround was to create a small wrapper struct that lifts the <K,V> to
the struct, and we implement `KeyValStore` on `LMDBWrapper` instead, which
works fine, but required quite a few changes across the codebase filling
this wrapper in.

If you can think of a better way to do this, I'm all ears :)