Add Mutable MMR Datastructure
Unlike a pure MMR, which is append-only, in `MutableMmr`, leaf nodes can be marked as deleted.
In `MutableMmr` a roaring bitmap tracks which data have been marked as deleted, and the merklish root is modified
to include the hash of the roaring bitmap.
The `MutableMmr` API maps nearly 1:1 to that of MerkleMountainRange so that you should be able to use it as a
drop-in replacement for the latter in most cases.