Port MerkleChangeTracker to use arbitrary backend (#680)
* Port MerkleChangeTracker to use arbitrary backend

This PR ports SWvHeerden's ChangeTracker to the new MMR crate to enable it to
be used with arbitrary backends and backing MMRs.  For example, you can build a
rewindable MMR with a pruned MMR based in LMDB with a memory-babcked
checkpointed MMR, or any combination of the two.

This PR adds the `MerkleChangeChecker` struct and supporting functions.  The
`ArrayLike` like was extended to include two new functions, `trunacate` and an
iterator-like function `for_each`.  For implementations that don't want to, or
can't support these new features, a `NotSupported` error can be returned.