Blockchain database: Refactor ProofOfWork
In preparation for a backend-agnostic BC-DB, we refactor the proof of work slightly.

The concept of a target and difficulty is laid out, and a demo PoW based on Blake2b is provided
that's slightly more functional than the existing `MockPoW`, including a simple
mining algorithm.

The current PoW modules remains in place for now -- an upcoming PR will replace it.