PR: Add `set_pruning_horizon` to blockchain database transaction
The pruning horizon is an important and general piece of metadata that informs
users of BlockchainDatabases where full storage of blockchain history begins.

The pruning horizon is used in synchronisation requests and block data requests
(e.g. from explorers).

Given that this is the 3rd piece of metadata, the fetch method is DRYd up to
eliminate a bunch of repeated code.

Changes summary:
* Cleans up `read_metadata` by using a macro to clean out boilerplate
* Allows the setting of a pruning horizon. BAckend can use this as a
  hint to delete older data
* Clean up some compiler warnings