Fetch historical blocks from Blockchain database (#714)
* Add `add_block` and `fetch_block` to `BlockchainDB`

The primary additions in the PR are the inclusion of `add_block` and
`fetch_block` to the BlockchainDatabase struct. This involved a few other minor
additions and breaking changes:

- `add_block` returns a `BlockResult`. This allows us to differentiate between
  a block being added to the longest chain, the orphan pool, etc.
- If a value doesn't exisst in the DB, `fetch_xxx` now returns a
  `ChainStorageError` variant rather than Ok(None). This change saves a ton of
boilerplate.
- Added a block hash of the chain tip to the metadata
- Added a block hash index to the database backend interface. Querying by block
  has returns the block height of the block with the given hash.
- Changed blockheader hash type to Vec<u8>, rater than [0; 32]. This was a
  tough choice. The DB backends are going to be returning Vec<u8>; and so there
isn't a super-cheap or concise way of converting between vectors and arrays,
because Vecs live on the heap and arrays live on the stack. So there are always
going to be copies happening. We now lose static bounds checks, but I'm not
sure that's useful when you're getting vecs from the DB backend anyway.
- Fix bug in `get_genesis_block()` that wasn't sorting entries
- Add `commitBlock` DbKey that gives backends a change to do cleanup after
  adding block data. Currently used to commit MMR checkpoints.
- Clean up Blocks module API; making everything private by default and exposing
  key structs to the top level.

* Add clarifications for review comments