Added tari_test_utils crate (#717)
Description

This crate will provide common utilities for testing the Tari base_layer codebase.

This should contain common utilities that are often needed for testing. Putting larger
code snippets which, for instance, set up a graph of network nodes or populate and
return an LMDB database with data specific to a single module should be avoided
as this could create circular dependencies/prevent crates from using these utilities.

Currently, this crate contains a utility macro (counter_context!) which
creates a future Context required when testing custom implementations of
the Future trait and a function to generate a random string of a given length.