Add reply_channel to comms
reply_channel is a simple wrapper around a mpsc unbounded channel
which implements `tower_service::Service` which, when called, includes
a oneshot for replies. This is exactly the same idea as the
reply_channel in service_framework, but simplified further and made
available in comms for future use. There does not appear to be a common
place to put this small wrapper which fits, and seems excessive to
create a new crate.

- Moved tari_test_utils into infrastructure so that comms can make use
  of it