Remove the need for ServiceName enum
The ServiceName enum was a way to index service handles so that they are
easily retrieved by key in the `ServiceHandles<ServiceName>` collection.
This meant that every crate which contains a service had to know about the
`ServiceName` enum and had to add a key to this enum.

To retreive a handle, you need the `ServiceName` enum and the type of
the handle.

```rust
let my_service = handles.get_handle::<MyServiceHandle>(ServiceName::MyService);
```

This change removes the need for the `ServiceName` enum. All that is
required to retreive a handle is the type for the handle.

```rust
let my_service = handles.get_handle::<MyServiceHandle>();
```

Internally this is done by using the `TypeId` for the type. Each static type is
assigned a globally unique u64 value. As long as the type for the
registered handle is unique per service, then there would be no
overwriting in the underlying HashMap.

Note that, as expected, generics are different types

```rust
TypeId::of::<ServiceSender<MyRequest, MyResponse>>() != TypeId::of::<SenderService<OtherRequest, OtherResponse>>()
```

so this should never be an issue.