Added "actor" for connection manager
The `ConnectionManagerActor` receives and furnishes requests to create connections
(a.k.a dial) to peers. Requests are made using an mpsc channel and are
processed concurrently and asynchonously.

This is an async front for a ConnectionManager (which is still
synchronous). The OMS will use this to establish connections and allow
the implementation to become fully async down the line.

To illustrate (does not compile):

```rust
// Create a requester/actor pair
let (requester, actor) = create_connection_manager_actor(/* connection manager goes here */);
// Spawn actor task which will respond to requests from the requester
tokio::spawn(actor.start());

// Request to dial `node_id`
tokio::block_on(async move {
  if let Ok(peer_connection) = requester.dial_node(node_id).await {
    println!("Success!")
  }
});
```

The Dialer trait is an abstraction that allows decoupling and ease of
testing. The ConnectionManager implements Dialer.