Outline of base node implementation
This PR simply sketches the outline and strategy for the Tari Base
node (aka full node) implementation. In summary, we describe the
node as a FSM; that maintains blockchain state with its peers.

This PR comprises mainly documentation.
Implied in the FSM is that the node contains validation services
that it delegates to in order to validate blocks as they are received.