Update the Wallet container and reintroduce it to the applications
This PR updates the Waller container object to support the new async services. 
Integration tests were added for the current Wallet services such as the wallet itself, TMS and the Liveness service.
The new wallet object was then plumbed into the Console Text Messenger application.

The gRPC wallet was not updated as it is not due for MVP and the FFI bindings still need to be updated to use a Tokio runtime.