Update OutputManagerService and TransactionService to be Async
This PR updates the Output manager and Transaction service in the Wallet crate to be fully async using the Service Framework. We also add an event stream to the transaction service and a variety of integration tests.

Following this work we were able to remove all the sync service code from the P2P crate which is no longer used anywhere in the codebase.