Convert Base Node state machine to use futures
The base node FSM is now futures-based. `run()` consumes self and returns
a Future that should be spawned onto a convenient runtime.

-- Shared state --

The `BaseNodeStateMachine` struct now holds shared state for the FSM,
which is passed into each state instance's `next_event` method.
We're sharing the DB, config, and the outgoing comms interface.

This simplifies the definitions of the individual state structs
considerably (bye-bye generics!)

-- InitialSync --

The InitialSync stae logic is now complete, albeit minimalistic.
In future, we'd like to add some heuristics for determining the
network consensus for the chain height, but for now, we just
use the maximum value given to us.

We also give up too easily. There should be a general automatic
retry-with-backoff utility in the FSM that all states can employ.
(TODO)

-- Miscellaneous changes: --

Introduce BaseNodeError

Generalise `Chainmetadata::horizon_block()` to calculate the horizon
block given any block height and the instance's pruning horizon; and
propagate this change throught the codebase.

Ran `cargo fix -p tari_core` to great effect!